Initial commit
This is the initial baseline for the project. Its derived from the SkiService project and has been modified to fit the requirements of this project.

Backend source: https://github.com/fokklz/165-ski-service-management-no-sql
The generics have been modified to allow for overriding the roles needed for a endpoint.
This is done trough the use of the CustomAuthorizeAttribute and the CustomAuthorizationPolicyProvider.

Model source: https://github.com/fokklz/nuget-ski-service-models
The general structure is the same, the complexity of the project has been reduced.

Application source: https://github.com/fokklz/ict-322-ski-service-app
There was not much work done on the application, the general structure is the same.
Core services have been added, the frontend will differ from the original project so only architecture is taken from this project.